+++ src/components/entries/Select.js
@@ -13,6 +13,7 @@ import {
 } from 'preact/hooks';
 
 import Description from './Description';
+import ErrorMessage from './Error';
 import Tooltip from './Tooltip';
 
 { /* Required to break up imports, see https://github.com/babel/babel/issues/15156 */ }
@@ -195,7 +196,7 @@ export default function SelectEntry(props) {
         disabled={ disabled }
         tooltip={ tooltip }
         element={ element } />
-      { error && <div class="bio-properties-panel-error">{ error }</div> }
+      <ErrorMessage error={ error } forId={ id } element={ element } />
       <Description forId={ id } element={ element } value={ description } />
     </div>
   );