+++ src/components/entries/TextField.js
@@ -1,4 +1,5 @@
 import Description from './Description';
+import ErrorMessage from './Error';
 import Tooltip from './Tooltip';
 
 import {
@@ -226,7 +227,7 @@ export default function TextfieldEntry(props) {
         value={ value }
         tooltip={ tooltip }
         element={ element } />
-      { error && <div class="bio-properties-panel-error">{ error }</div> }
+      <ErrorMessage error={ error } forId={ id } element={ element } />
       <Description forId={ id } element={ element } value={ description } />
     </div>
   );